+++ dj/project/asgi.py
@@ -1,6 +1,6 @@
 import os
 from django.core.asgi import get_asgi_application
 
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'djplus.project.settings')
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'dj.project.settings.dev')
 
 application = get_asgi_application()