+++ tests/auth/__init__.py
@@ -1,6 +1,6 @@
 from django.core.signals import setting_changed
 from django.dispatch import receiver
-from djplus.auth.hashers import get_hashers
+from dj.auth.hashers import get_hashers
 
 
 @receiver(setting_changed)