+++ tests/blog/test_managers.py
@@ -1,7 +1,7 @@
 from django.test import TestCase
 from django.utils import timezone
-from djplus.auth.models import User
-from djplus.blog.models import Post
+from dj.auth.models import User
+from dj.blog.models import Post
 
 
 class PublishedManagerTests(TestCase):
@@ -13,6 +13,6 @@ class PublishedManagerTests(TestCase):
         Post.objects.create(title="What is Django?", slug="what-is-django", author=self.user)
         Post.objects.create(title="What is Python?", slug="what-is-python", author=self.user,
                             publication_date=timezone.now()+timezone.timedelta(seconds=1))
-        post = Post.objects.create(title="What is Djplus?", slug="what-is-djplus", author=self.user,
+        post = Post.objects.create(title="What is Djplus?", slug="what-is-dj", author=self.user,
                                    publication_date=timezone.now() - timezone.timedelta(seconds=1))
         self.assertEqual(list(Post.published.all()), [post])