+++ tests/blog/test_views.py
@@ -1,10 +1,10 @@
 from django.test import TestCase, override_settings
 from django.utils import timezone
-from djplus.blog.models import Post
-from djplus.auth.models import User
+from dj.blog.models import Post
+from dj.auth.models import User
 
 
-@override_settings(ROOT_URLCONF="djplus.blog.urls")
+@override_settings(ROOT_URLCONF="dj.blog.urls")
 class IndexViewTests(TestCase):
     def test_template_name(self):
         response = self.client.get("/")
@@ -15,7 +15,7 @@ class IndexViewTests(TestCase):
         self.assertEqual(response.status_code, 200)
 
 
-@override_settings(ROOT_URLCONF="djplus.blog.urls")
+@override_settings(ROOT_URLCONF="dj.blog.urls")
 class PostDetailViewTests(TestCase):
     @classmethod
     def setUpTestData(cls):
@@ -38,9 +38,9 @@ class PostDetailViewTests(TestCase):
 
     def test_show_only_published_post(self):
         Post.objects.create(title="What is Django?", slug="what-is-django", author=self.user)
-        Post.objects.create(title="What is Djplus?", slug="what-is-djplus", author=self.user,
+        Post.objects.create(title="What is Djplus?", slug="what-is-dj", author=self.user,
                             publication_date=timezone.now() + timezone.timedelta(seconds=1))
         response = self.client.get("/what-is-django/")
         self.assertEqual(response.status_code, 404)
-        response = self.client.get("/what-is-djplus/")
+        response = self.client.get("/what-is-dj/")
         self.assertEqual(response.status_code, 404)