+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add initial djplus project.
 - Add `ip` field to `Session` model.
 - Add `status` field to `User` model.
+### Breaking Changes
+- Change the name of the package from `djplus` to `dj`.
 
 ## [2.1.0] - 2022-10-27
 ### Added